Add files via upload
This problem has BasePrice, EngineChoice, TrimChoice, and RadioChoice as inputs and SellingPrice as the only output.
My program must 1)input the base price of car, 2)Process the various option choices to compute additional costs, 3) Total all the costs, and 4) Display the final selling price for the car and choices that you have made.

If it is at all possible try to explain what I was doing wrong or explain what new steps you added! I know this is a lot but Im desperate and my professor doesnt help at all!